requirements.txt
selenium
webdriver-manager
pytest
pytest-xdist==2.2.0
python-dotenv
pytest-html
assertpy
openpyxl